+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## New Features
 
+- Add `api::SystemResource` to encapsulate system resources and minimize the use of the rule of 5.
 - Documentation updates for `API`
 - Use a `printer::Printer*` to print help when using `sys::Cli::show_help()`
 - Add `sys::System::get_executable_suffix()` for desktop builds